backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pandas
scikit-learn==1.4.2
psycopg2-binary
joblib
passlib[bcrypt]
bcrypt==3.2.0
python-multipart
python-jose[cryptography]
openai
python-dotenv
